+++ src/client/src/components/styledComponents.ts
@@ -1,4 +1,5 @@
 import { Box, Stack } from "@mui/material";
+import ListItem from "@mui/material/ListItem";
 import { styled } from "@mui/system";
 import { theme } from "../AppTheme";
 
@@ -69,3 +70,28 @@ export const DetailHeaderStack = styled(Stack)({
   height: "84px",
   padding: "16px",
 });
+
+interface ParentListItemProps {
+  active: boolean;
+}
+
+export const ParentListItem = styled(ListItem, {
+  shouldForwardProp: prop => prop !== "active", // Prevents "active" from being forwarded as a DOM attribute
+})<ParentListItemProps>(({ active }) => ({
+  padding: "1em",
+  display: "flex",
+  height: "40px",
+  cursor: "pointer",
+  paddingLeft: "35.5px",
+  color: active ? theme.palette.error.main : "",
+  borderTop: `1px solid ${theme.palette.border.light}`,
+  borderLeft: active ? `0.25em solid ${theme.palette.error.main}` : undefined,
+  backgroundColor: active ? theme.palette.background.lightgrey : "",
+  "&:hover": {
+    backgroundColor: theme.palette.hover.main,
+  },
+}));
+
+export const ChildListItem = styled(ParentListItem)(() => ({
+  paddingLeft: "50px !important",
+}));