+++ src/client/src/pages/detail/detailHeaderSettings.tsx
@@ -1,10 +1,13 @@
+import { useTranslation } from "react-i18next";
 import { useHistory } from "react-router-dom";
-import { Stack } from "@mui/material";
+import { Stack, Typography } from "@mui/material";
 import { ReturnButton } from "../../components/buttons/buttons.tsx";
 import { DetailHeaderStack } from "../../components/styledComponents.ts";
+import { capitalizeFirstLetter } from "../../utils.ts";
 
 export const DetailHeaderSettings = () => {
   const history = useHistory();
+  const { t } = useTranslation();
 
   return (
     <DetailHeaderStack direction="row" alignItems="center">
@@ -14,6 +17,7 @@ export const DetailHeaderSettings = () => {
             history.push("/");
           }}
         />
+        <Typography variant="h2"> {capitalizeFirstLetter(t("settings"))}</Typography>
       </Stack>
     </DetailHeaderStack>
   );