+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Updated the UI design of the tabs component.
 - Reincluded download link for codelist references in the import panel.
 - Formfield border color now depend on whether borehole editing is enabled or not.
+- The settings are now displayed in tabs.
 
 ### Fixed